+++ src/components/Shared/Modal/GameModal.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import PropTypes from "prop-types";
+
+const GameModal = () => {};
+
+export default GameModal;
