+++ src/components/Shared/Modal/GameModalStyle.jsx
@@ -1,0 +1,2 @@
+import styled from "styled-components";
+import { FlexBox } from "@/styles/FlexStyle";
