+++ src/components/Shared/Modal/ProfileModal.jsx
@@ -1,0 +1,6 @@
+import React from "react";
+import PropTypes from "prop-types";
+
+const ProfileModal = () => {};
+
+export default ProfileModal;
