+++ src/components/Shared/Modal/ProfileModalStyle.jsx
@@ -1,0 +1,2 @@
+import styled from "styled-components";
+import { FlexBox } from "@/styles/FlexStyle";
